part 6 a
